Slider now have Live update of budget ratio, Slider will be at the selected value of the user when loaded in from a saved event

+++ src/HotelSuggest.java
@@ -56,7 +56,7 @@ public class HotelSuggest {
 		this.eventBudget = eventBudget;
 	}
 
-	public double getBudgetRatio() {
+	public int getBudgetRatio() {
 		return budgetRatio;
 	}
